Use node v18 and net 6 and 7.

+++ pack.sh
@@ -15,8 +15,16 @@ SRC_DIR=$(dirname $(readlink -f "$0"))
 DST_DIR="$SRC_DIR/packed"
 TEMP_DIR=$(mktemp -d)
 
-# setup auto cleanup
-trap "rm -rf $TEMP_DIR" EXIT
+if [[ "$2" = "--keep" ]]; then
+    KEEP="true"
+else
+    KEEP="false"
+fi
+
+if [[ "$KEEP" = "false" ]]; then
+    # setup auto cleanup
+    trap "rm -rf $TEMP_DIR" EXIT
+fi
 
 pushd "$TEMP_DIR"
 echo "Using temporary directory $TEMP_DIR"
@@ -29,10 +37,10 @@ yarn install
 yarn add "prettier@$PRETTIER_VERSION" @prettier/plugin-xml prettier-plugin-sh
 
 mkdir -p "$DST_DIR/win-x64/"
-yarn pkg -t node14-win ./node_modules/prettier/bin-prettier.js -o "$DST_DIR/win-x64/prettier.exe" --config ./pkg.config.json
+yarn pkg -t node18-win ./node_modules/prettier/bin-prettier.js -o "$DST_DIR/win-x64/prettier.exe" --config ./pkg.config.json
 
 mkdir -p "$DST_DIR/linux-x64/"
-yarn pkg -t node14-linux ./node_modules/prettier/bin-prettier.js -o "$DST_DIR/linux-x64/prettier" --config ./pkg.config.json
+yarn pkg -t node18-linux ./node_modules/prettier/bin-prettier.js -o "$DST_DIR/linux-x64/prettier" --config ./pkg.config.json
 
 echo -n $PRETTIER_VERSION > "$DST_DIR/version"
